merge: move descriptions.py and core.py to data/
